# [OpenCL] Optimized Single-Precision GEMM Kernel