This pull request adds a highly optimized single-precision General Matrix Multiplication (GEMM) kernel developed for OpenCL. The enhancements in this kernel aim to improve computational efficiency and performance for matrix operations, reducing execution time and enhancing throughput.

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped

## Result

#### PC (x86)
|engine|avg|
|------|---|
|CPU|108.05ms|
|GPU (prev)|3848.6 ms|
|GPU|68.5 ms|

#### Android (aarch64)
|engine|avg|
|------|---|
|CPU|293.75ms|
|GPU (prev)|19600.8 ms|
|GPU|534.1 ms|

## Note

This is a profiling result with the following matrix size.
- M: 1024
- K: 3072
- N: 3072