@djeong20 2025-04-21
> So this kernel aims to store 16x16 ukernel computed with (16,16)x(16x16) sub-A and sub-B block? Looks good to me overall, but I have some ideas to test further🤔

Yes, this utilizes local memory for matrices A and B. Please feel free to share ideas!