@djeong20 2025-05-08
> LGTM.
> 
> How can I test?

You can check out this PR and modify the unittest blas kernel. Just so you know, I used the chrono library, ran 20 times, and averaged it.